requirements.txt
annotated-types==0.7.0
anyio==4.8.0
appscript==1.3.0
beautifulsoup4==4.12.3
boto3==1.36.11
botocore==1.36.11
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
duckduckgo_search==7.2.1
feedparser==6.0.11
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.7
httpx==0.28.1
icalendar==6.1.1
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jiter==0.8.2
jmespath==1.0.1
keyring==24.1.1
keyrings.alt==4.2.0
lxml==5.3.0
more-itertools==10.6.0
openai==1.61.1
-e git+https://github.com/smankoo/personal-assistant.git@82e0ad2266016fced4692621089647a9b516d731#egg=personal_assistant
primp==0.10.1
pydantic==2.10.6
pydantic_core==2.27.2
pyicloud==1.0.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.11.2
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
srp==1.0.22
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.1
urllib3==2.3.0
